Refactor CLI code composition
