+++ lib/rage/cli.rb
@@ -4,65 +4,13 @@ require "thor"
 require "rack"
 require "rage/version"
 
+require "rage/cli/base"
 require "rage/cli/skills"
+require "rage/cli/code_generator"
+require "rage/cli/new_app_generator"
 
-module Rage
-  class CLICodeGenerator < Thor
-    include Thor::Actions
-
-    def self.source_root
-      File.expand_path("templates", __dir__)
-    end
-
-    desc "migration NAME", "Generate a new migration"
-    def migration(name = nil)
-      return help("migration") if name.nil?
-
-      setup
-      Rake::Task["db:new_migration"].invoke(name)
-    end
-
-    desc "model NAME", "Generate a new model"
-    def model(name = nil)
-      return help("model") if name.nil?
-
-      setup
-      migration("create_#{name.pluralize}")
-      @model_name = name.classify
-      template("model-template/model.rb", "app/models/#{name.singularize.underscore}.rb")
-    end
-
-    desc "controller NAME", "Generate a new controller"
-    def controller(name = nil)
-      return help("controller") if name.nil?
-
-      setup
-      unless defined?(ActiveSupport::Inflector)
-        raise LoadError, <<~ERR
-          ActiveSupport::Inflector is required to run this command. Add the following line to your Gemfile:
-          gem "activesupport", require: "active_support/inflector"
-        ERR
-      end
-
-      # remove trailing Controller if already present
-      normalized_name = name.sub(/_?controller$/i, "")
-      @controller_name = "#{normalized_name.camelize}Controller"
-      file_name = "#{normalized_name.underscore}_controller.rb"
-
-      template("controller-template/controller.rb", "app/controllers/#{file_name}")
-    end
-
-    private
-
-    def setup
-      @setup ||= begin
-        require "rake"
-        load "Rakefile"
-      end
-    end
-  end
-
-  class CLI < Thor
+module Rage::CLI
+  class App < Base
     def self.exit_on_failure?
       true
     end
@@ -74,7 +22,7 @@ module Rage
       return help("new") if options.help? || path.nil?
 
       require "rage/all"
-      CLINewAppGenerator.start([path, options[:database]])
+      NewAppGenerator.start([path, options[:database]])
     end
 
     desc "s", "Start the app server"
@@ -232,11 +180,11 @@ module Rage
     end
 
     desc "skills", "Manage coding agent skills"
-    subcommand "skills", CLISkills
+    subcommand "skills", Skills
 
     map "generate" => :g
     desc "g TYPE", "Generate new code"
-    subcommand "g", CLICodeGenerator
+    subcommand "g", CodeGenerator
 
     map "--tasks" => :tasks
     desc "--tasks", "See the list of available tasks"
@@ -272,26 +220,6 @@ module Rage
     end
 
     private
-
-    def environment
-      require File.expand_path("config/application.rb", Dir.pwd)
-
-      if Rage.config.internal.rails_mode
-        require File.expand_path("config/environment.rb", Dir.pwd)
-      end
-    end
-
-    def set_env(options)
-      if options[:environment]
-        ENV["RAGE_ENV"] = ENV["RAILS_ENV"] = options[:environment]
-      elsif ENV["RAGE_ENV"]
-        ENV["RAILS_ENV"] = ENV["RAGE_ENV"]
-      elsif ENV["RAILS_ENV"]
-        ENV["RAGE_ENV"] = ENV["RAILS_ENV"]
-      else
-        ENV["RAGE_ENV"] = ENV["RAILS_ENV"] = "development"
-      end
-    end
 
     def linked_rake_tasks
       require "rake"
@@ -368,69 +296,6 @@ module Rage
           prefix = subscriber == subscribers.last && next_level != level + 1 ? "└─" : "├─"
           puts "│#{padding * (level + 1)}#{prefix} \e[1m#{subscriber}\e[0m"
         end
-      end
-    end
-  end
-
-  class CLINewAppGenerator < Thor::Group
-    include Thor::Actions
-    argument :path, type: :string
-    argument :database, type: :string, required: false
-
-    def self.source_root
-      File.expand_path("templates", __dir__)
-    end
-
-    def setup
-      @use_database = !database.nil?
-    end
-
-    def create_directory
-      empty_directory(path)
-    end
-
-    def copy_files
-      inject_templates
-    end
-
-    def install_database
-      return unless @use_database
-
-      @app_name = path.tr("-", "_").downcase
-      append_to_file "#{path}/Gemfile", <<~RUBY
-
-        gem "#{get_db_gem_name}"
-        gem "activerecord"
-        gem "standalone_migrations", require: false
-      RUBY
-
-      inject_templates("db-templates")
-      inject_templates("db-templates/#{database}")
-    end
-
-    private
-
-    def inject_templates(from = nil)
-      root = "#{self.class.source_root}/#{from}"
-
-      Dir.glob("*", base: root).each do |template|
-        next if File.directory?("#{root}/#{template}")
-
-        *template_path_parts, template_name = template.split("-")
-        template("#{root}/#{template}", [path, *template_path_parts, template_name].join("/"))
-      end
-    end
-
-    def get_db_gem_name
-      case database
-      when "mysql"
-        "mysql2"
-      when "trilogy"
-        "trilogy"
-      when "postgresql"
-        "pg"
-      when "sqlite3"
-        "sqlite3"
       end
     end
   end
